// src/server/types/openai/response.ts
// Generated by https://quicktype.io

export interface OpenAICompletionResponse {
    id:      string;
    object:  string;
    created: number;
    model:   string;
    choices: Choice[];
    usage:   Usage;
}

export interface Choice {
    text:          string;
    index:         number;
    logprobs:      null;
    finish_reason: string;
}

export interface Usage {
    prompt_tokens:     number;
    completion_tokens: number;
    total_tokens:      number;
}